Force an update of a sample's LCMethod everywhere we really need a clone, or to unlink it from the config/setup time LCMethods
This fixes to progress images and the test run event counts, meaning everything appears to work as it should.
I tried to instead add a "ActualLCMethod" property to classSampleData, and use it everywhere that I could determine that a sample runtime copy was needed, but wasn't able to get everything to work properly again.