Housekeeping: cleaning up project references and post-build event changes
Removed unused references
Creating a documentation file for all DLLs, but ignoring the warning for public properties/methods/classes that don't have XML documentation
Copying the DLL, .pdb, and .xml doc file for each project as part of the post-build events, but not copying the referenced DLLs unless they are specific to that project (besides NuGet packages). This is to clean up the lib and pluginDLLs directories, and avoid DLL issues when debugging or otherwise. Specifically in this case, the FluidicsSDK in the Plugins folder would get loaded, and would cause problems because the debugger would try to use the one in the same folder as the main exe, and throw an error that would block breakpoints and stepping into the FluidicsSDK code. The DLL loader will always include the directory containing the executable in the DLL resolution process, so it's smarter to only have one copy of each DLL among the install and DLL loading directories.