Add support for method events that can end early without bad side effects
Previous code would wait for the timeout to end, to properly handle multi-column synchronization.
Code changes allow moving on to another event without waiting for the timeout, in the following circumstances:
1) Event has an attribute that permits moving on before the time out
2) Only a single column is enabled, OR the running method does not allow Pre/Post overlap

When an event ends early, the duration and following event start times are reset, so that they do not add unexpected delays later in the method. Also, when moving on to subsequent events, there is a similar check to see if there is an unreasonable amount of time before the next method is scheduled to start; if there is, and the method does not allow Pre/Post overlap, then the start time for the next event will be adjusted to ~10 seconds after the first check.