Remove several plugins that are no longer used (or potentially useful) for current LC setups, and their respective settings from LCMSNetProg
This also removes some otherwise-unused settings.
Some of the plugins removed were never completed/functional.
The rest of the plugins removed have not been used in at least 5 years, with the setups that used them since dismantled, and no desire to create a new setup that would use that plugin.