Add some flexibility to the method loading, mostly to support reasonable changes to hardware config
Instead of always storing the device type name, store the name of the type that implements the method used by the event. This is usually the same as the device type name, but for some devices the only difference from another device in code is configuration details, like the number of ports on the 2-position valve.

A common example of a reasonable change: changing the displayed type of a six-port valve, between 'injection with loop', six-port SPE, and standard six-port valve.